LWDEV-6992 use big decimal in services

+++ src/main/kotlin/com/lykke/matching/engine/daos/HourCandle.kt
@@ -1,10 +1,11 @@
 package com.lykke.matching.engine.daos
 
+import java.math.BigDecimal
 import java.util.LinkedList
 
-class HourCandle(val asset: String, val prices: LinkedList<Double>) {
+class HourCandle(val asset: String, val prices: LinkedList<BigDecimal>) {
 
-    fun addPrice(price: Double?) {
+    fun addPrice(price: BigDecimal?) {
         if (price != null) {
             prices.add(price)
             while (prices.size < 20) {
