+++ src/main/kotlin/com/lykke/matching/engine/daos/balance/ReservedVolumeCorrection.kt
@@ -1,7 +1,9 @@
 package com.lykke.matching.engine.daos.balance
 
+import java.math.BigDecimal
+
 data class ReservedVolumeCorrection(val clientId: String,
                                     val assetId: String,
                                     val orderIds: String?,
-                                    val oldReserved: Double,
-                                    val newReserved: Double)
+                                    val oldReserved: BigDecimal,
+                                    val newReserved: BigDecimal)
