+++ src/main/kotlin/com/lykke/matching/engine/database/azure/AzureCashOperationsDatabaseAccessor.kt
@@ -20,7 +20,8 @@ class AzureCashOperationsDatabaseAccessor (connectionString: String,
 
     override fun insertTransferOperation(operation: TransferOperation) {
         try {
-            transferOperationsTable.execute(TableOperation.insertOrMerge(AzureWalletTransferOperation(operation.id, operation.externalId, operation.fromClientId, operation.toClientId, operation.asset, operation.dateTime, operation.volume)))
+            transferOperationsTable.execute(TableOperation.insertOrMerge(AzureWalletTransferOperation(operation.id, operation.externalId,
+                    operation.fromClientId, operation.toClientId, operation.asset, operation.dateTime, operation.volume.toDouble())))
         } catch (e: Exception) {
             LOGGER.error("Unable to insert operation: ${operation.id}, external id: ${operation.externalId}", e)
             METRICS_LOGGER.logError("Unable to insert operation: ${operation.id}, external id: ${operation.externalId}", e)
